requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
pymysql==1.1.0
flask-jwt-extended==4.6.0
cryptography==41.0.7
Werkzeug==3.0.1
pyyaml==6.0.1
